chore: move files to new view
