+++ app/components/UI/Predict/views/PredictBuyWithAnyToken/PredictBuyWithAnyToken.tsx
@@ -19,25 +19,25 @@ import { useSelector } from 'react-redux';
 import { BottomSheetRef } from '../../../../../component-library/components/BottomSheets/BottomSheet';
 import { TraceName } from '../../../../../util/trace';
 import { PredictBuyPreviewSelectorsIDs } from '../../Predict.testIds';
-import PredictBuyActionButton from '../../components/PredictBuyActionButton';
-import PredictBuyAmountSection from '../../components/PredictBuyAmountSection';
-import PredictBuyBottomContent from '../../components/PredictBuyBottomContent';
-import PredictBuyMinimumError from '../../components/PredictBuyMinimumError';
-import PredictBuyPreviewHeader from '../../components/PredictBuyPreviewHeader/PredictBuyPreviewHeader';
+import PredictBuyActionButton from './components/PredictBuyActionButton';
+import PredictBuyAmountSection from './components/PredictBuyAmountSection';
+import PredictBuyBottomContent from './components/PredictBuyBottomContent';
+import PredictBuyMinimumError from './components/PredictBuyMinimumError';
+import PredictBuyPreviewHeader from './components/PredictBuyPreviewHeader/PredictBuyPreviewHeader';
 import PredictFeeBreakdownSheet from '../../components/PredictFeeBreakdownSheet';
-import PredictFeeSummary from '../../components/PredictFeeSummary/PredictFeeSummary';
+import PredictFeeSummary from './components/PredictFeeSummary/PredictFeeSummary';
 import PredictKeypad, {
   PredictKeypadHandles,
 } from '../../components/PredictKeypad';
 import PredictOrderRetrySheet from '../../components/PredictOrderRetrySheet';
-import PredictPayWithAnyTokenInfo from '../../components/PredictPayWithAnyTokenInfo';
-import { PredictPayWithRow } from '../../components/PredictPayWithRow';
-import { usePredictBuyAvailableBalance } from '../../hooks/usePredictBuyAvailableBalance';
-import usePredictBuyBackSwipe from '../../hooks/usePredictBuyBackSwipe';
-import { usePredictBuyConditions } from '../../hooks/usePredictBuyConditions';
-import { usePredictBuyInfo } from '../../hooks/usePredictBuyInfo';
-import { usePredictBuyInputState } from '../../hooks/usePredictBuyInputState';
-import { usePredictBuyActions } from '../../hooks/usePredictBuyPreviewActions';
+import PredictPayWithAnyTokenInfo from './components/PredictPayWithAnyTokenInfo';
+import { PredictPayWithRow } from './components/PredictPayWithRow';
+import { usePredictBuyAvailableBalance } from './hooks/usePredictBuyAvailableBalance';
+import usePredictBuyBackSwipe from './hooks/usePredictBuyBackSwipe';
+import { usePredictBuyConditions } from './hooks/usePredictBuyConditions';
+import { usePredictBuyInfo } from './hooks/usePredictBuyInfo';
+import { usePredictBuyInputState } from './hooks/usePredictBuyInputState';
+import { usePredictBuyActions } from './hooks/usePredictBuyPreviewActions';
 import { usePredictMeasurement } from '../../hooks/usePredictMeasurement';
 import { usePredictOrderPreview } from '../../hooks/usePredictOrderPreview';
 import { usePredictOrderRetry } from '../../hooks/usePredictOrderRetry';
