+++ app/components/UI/Predict/views/PredictBuyWithAnyToken/components/PredictBuyActionButton/index.ts
