+++ app/components/UI/Predict/views/PredictBuyWithAnyToken/components/PredictBuyAmountSection/index.ts
