+++ app/components/UI/Predict/views/PredictBuyWithAnyToken/components/PredictBuyBottomContent/index.ts
