+++ app/components/UI/Predict/views/PredictBuyWithAnyToken/components/PredictBuyMinimumError/index.ts
