+++ app/components/UI/Predict/views/PredictBuyWithAnyToken/components/PredictBuyPreviewHeader/index.ts
