+++ app/components/UI/Predict/views/PredictBuyWithAnyToken/components/PredictPayWithRow/index.ts
